app/src/main/java/com/englishlearn/myapplication/data/TractateType.java
package com.englishlearn.myapplication.data;

import java.io.Serializable;

/**
 * Created by yanzl on 16-9-3.
 * 信息分类表
 */
public class TractateType implements Serializable,Cloneable{

    private String objectId;
    private String name;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getObjectId() {
        return objectId;
    }

    public void setObjectId(String objectId) {
        this.objectId = objectId;
    }

    @Override
    public Object clone(){

        Object o = null;
        try {
            o = super.clone();
        } catch (CloneNotSupportedException e) {
            e.printStackTrace();
        }
        return o;
    }

    @Override
    public String toString() {
        return "TractateType{" +
                "objectId='" + objectId + '\'' +
                ", name='" + name + '\'' +
                '}';
    }
}
